Add useQuerys for today's reviews
