+++ src/lib/use-reviewables.ts
@@ -1,33 +1,10 @@
 import type { DailyCacheKey, ReviewStages } from '@/types/main'
-import supabase from './supabase-client'
-import { pids, ReviewInsert, ReviewRow, ReviewUpdate, uuid } from '@/types/main'
+import { pids, ReviewRow, uuid } from '@/types/main'
 import {
 	QueryClient,
 	queryOptions,
-	useMutation,
-	useQueryClient,
 	useSuspenseQuery,
 } from '@tanstack/react-query'
-import toast from 'react-hot-toast'
-import { PostgrestError } from '@supabase/supabase-js'
-
-const postReview = async (submitData: ReviewInsert) => {
-	const { data } = await supabase
-		.rpc('insert_user_card_review', submitData)
-		.throwOnError()
-
-	return data
-}
-
-const updateReview = async (submitData: ReviewUpdate) => {
-	if (!submitData?.review_id || !submitData?.score)
-		throw new Error('Invalid inputs; cannot update')
-	const { data } = await supabase
-		.rpc('update_user_card_review', submitData)
-		.throwOnError()
-
-	return data
-}
 
 function getFromLocalStorage<T>(queryKey: DailyCacheKey): null | T {
 	const data = localStorage.getItem(JSON.stringify(queryKey))
@@ -153,14 +130,6 @@ export const setReviewStage = (
 	})
 }
 
-export const useOneReview = (dailyCacheKey: DailyCacheKey, pid: uuid) =>
-	useSuspenseQuery({
-		queryKey: [...dailyCacheKey, pid],
-		queryFn: () => getReviewFromLocalStorage(dailyCacheKey, pid),
-		refetchOnMount: true,
-		refetchOnWindowFocus: true,
-	})
-
 // for when you DON'T want to send a mutation to the DB
 // (second review in a day doesn't count for the algo)
 export const setOneReview = (
@@ -171,69 +140,4 @@ export const setOneReview = (
 ) => {
 	queryClient.setQueryData([...dailyCacheKey, pid], data)
 	setFromLocalStorage([...dailyCacheKey, pid], data)
-}
-
-// for when you want to send a mutation to the DB
-// and update the local cache and this custom query cache
-export function useReviewMutation(
-	pid: uuid,
-	dailyCacheKey: DailyCacheKey,
-	proceed: () => void,
-	resetRevealCard: () => void
-) {
-	const queryClient = useQueryClient()
-	const { data: prevData } = useOneReview(dailyCacheKey, pid)
-	const { data: state } = useReviewState(dailyCacheKey)
-
-	return useMutation<ReviewRow, PostgrestError, { score: number }>({
-		mutationKey: [...dailyCacheKey, pid],
-		// @ts-expect-error ts-2322 -- because Supabase view fields are always | null
-		mutationFn: async ({ score }: { score: number }) => {
-			// We want 1 mutation per day per card. We can send a second mutation
-			// only during stage 1 or 2, to _correct_ an improper input.
-
-			// no mutations when re-reviewing incorrect
-			if (state.reviewStage > 3)
-				return {
-					...prevData,
-					score,
-				}
-
-			// during stages 1 and 2 send an update only if the score has changed
-			if (prevData?.score === score) return prevData
-			if (prevData?.id)
-				return await updateReview({
-					score,
-					review_id: prevData.id,
-				})
-
-			// standard case: card has not been reviewed today
-			return await postReview({
-				score,
-				phrase_id: pid,
-				lang: dailyCacheKey[1],
-				day_session: dailyCacheKey[3],
-			})
-		},
-		onSuccess: (data) => {
-			if (data.score === 1)
-				toast('okay', { icon: '🤔', position: 'bottom-center' })
-			if (data.score === 2)
-				toast('okay', { icon: '🤷', position: 'bottom-center' })
-			if (data.score === 3)
-				toast('got it', { icon: '👍️', position: 'bottom-center' })
-			if (data.score === 4) toast.success('nice', { position: 'bottom-center' })
-
-			const mergedData = { ...prevData, ...data }
-			setOneReview(dailyCacheKey, pid, mergedData, queryClient)
-			setTimeout(() => {
-				resetRevealCard()
-				proceed()
-			}, 1000)
-		},
-		onError: (error) => {
-			toast.error(`There was an error posting your review: ${error.message}`)
-			console.log(`Error posting review:`, error)
-		},
-	})
 }
